+++ libphx/src/LuaScheduler.rs
@@ -124,7 +124,7 @@ unsafe extern "C" fn LuaScheduler_Add(mut L: *mut Lua) -> libc::c_int {
         this.addQueue_size = this.addQueue_size + 1;
         *(this.addQueue_data).offset(fresh0 as isize) = elem;
     } else {
-        if (this.elems_capacity == this.elems_size) as libc::c_int as libc::c_long
+        if (this.elems_capacity == this.elems_size) as libc::c_long
             != 0
         {
             this
@@ -204,7 +204,7 @@ unsafe extern "C" fn LuaScheduler_Update(mut L: *mut Lua) -> libc::c_int {
         this.addQueue_size -= 1;
         let mut elem_0: SchedulerElem = *(this.addQueue_data)
             .offset(this.addQueue_size as isize);
-        if (this.elems_capacity == this.elems_size) as libc::c_int as libc::c_long
+        if (this.elems_capacity == this.elems_size) as libc::c_long
             != 0
         {
             this
